Create Finding_clumps_ofKmers_L_t.py
This program can be used to find clumps of k-mers. The patterns form clumps in the given dataset. 
The program has been written using Python. 
Input: A string Genome, and integers k, L, and t.
Output: All distinct k-mers forming (L, t)-clumps in Genome.
L = fixed length (short interval)  the genome.
k = k-mer as a “clump” if it appears many times within a short interval of the genome.

Given integers L and t, a k-mer Pattern forms an (L, t)-clump inside a (larger) string Genome if there is an interval of Genome of length L in which this k-mer appears at least t times.